+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.22](https://github.com/AaronConlon/codia/compare/v0.1.21...v0.1.22) (2026-07-16)
+
+### Features
+
+* add a pre-rendered try-it image and dark playground
+* animate try-it image refreshes
+
 ### [0.1.21](https://github.com/AaronConlon/codia/compare/v0.1.20...v0.1.21) (2026-07-14)